[WEB-104] added use pagination custom hook
